Backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-dotenv
google-genai
supabase
bcrypt==4.0.1
python-multipart